% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_file.R, R/load.R, R/sap_load.R
\name{isi_df}
\alias{isi_df}
\title{dataframe to graph}
\usage{
isi_df(file)

isi_df(file)

isi_df(file)
}
\arguments{
\item{file}{the isi file (txt) downloaded from Web of Science}

\item{isi_file}{isi dataframe format}
}
\value{
a graph object with the citation network

a dataframe with all variables in columns

a dataframe with all variables in columns
}
\description{
this function convert the dataframe file to graph

Load isi file (txt) and convert it to a data frame with a bibliometrix
function called convert2df().

Load isi file (txt) and convert it to a data frame with a bibliometrix
function called convert2df().
}
\examples{
\dontrun{
df <- sap::isi_df("scientometrics_1_1059.txt")
}
\dontrun{
df <- sap::isi_df("scientometrics_1_1059.txt")
}
}
\author{
Sebastian Robledo

Sebastian Robledo
}
