% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsap.R
\docType{package}
\name{rsap}
\alias{rsap}
\alias{rsap-package}
\title{sap: sap algorithm to scientometrics analysis}
\description{
Easily generate a new Tree of Science
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/coreofscience/r-sap/issues}
}

}
\author{
Sebastian Robledo \email{sebastian.robledogi@amigo.edu.co}
}
