% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_graph.R
\name{sap_graph}
\alias{sap_graph}
\title{Creating a graph object}
\usage{
sap_graph(isi_df)
}
\arguments{
\item{isi_df}{a dataframe with wos data}
}
\value{
a graph object
}
\description{
from df isi to graph
}
\details{
Load a dataframe with isi data and convert it
in a graph
}
\author{
Sebastian Robledo
}
