% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_load.R
\name{sap_load}
\alias{sap_load}
\title{Loading isi files}
\usage{
sap_load(file)
}
\arguments{
\item{file}{the isi file (txt) downloaded from Web of Science}
}
\value{
a dataframe with all variables in columns
}
\description{
load isi file and convert it to df
}
\details{
Load isi file (txt) and convert it to a data frame with a bibliometrix
function called convert2df().
}
\examples{
\dontrun{
df <- sap::sap_load("scientometrics_1_1059.txt")
}
}
\author{
Sebastian Robledo
}
