% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_process.R
\name{sap_process}
\alias{sap_process}
\title{Creating our ToS}
\usage{
sap_process(graph_file)
}
\arguments{
\item{graph_file}{a dataframe with wos data}
}
\value{
a graph object
}
\description{
from graph isi to sap - Tree of Science
}
\details{
Load a dataframe with isi data and convert it
in a graph
}
\author{
Sebastian Robledo
}
