% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_tos.R
\name{sap_tos}
\alias{sap_tos}
\title{sap}
\usage{
sap_tos(file)
}
\arguments{
\item{file}{the isi file (txt) downloaded from web of science}
}
\value{
a dataframe with roots, trunk, and leaves
}
\description{
sap algorithm
}
\details{
sap algorithm represents the metaphor of the
tree sap
}
\author{
Sebastian Robledo
}
