% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_dataset.R
\docType{data}
\name{wos}
\alias{wos}
\title{Web of Science Data}
\format{
a txt file 
\describe{
 \item{TI}{Title of the papers}
 }
 @source Web of Science
}
\usage{
wos
}
\description{
A dataset containing a search of Web of Science
}
\keyword{datasets}
